--- lib/binding/create_review_binding.dart
@@ -1,9 +1,0 @@
-import 'package:get/get.dart';
-import 'package:satorio/controller/create_review_controller.dart';
-
-class CreateReviewBinding extends Bindings {
-  @override
-  void dependencies() {
-    Get.lazyPut<CreateReviewController>(() => CreateReviewController());
-  }
-}
