+++ lib/binding/write_review_binding.dart
@@ -1,0 +1,9 @@
+import 'package:get/get.dart';
+import 'package:satorio/controller/write_review_controller.dart';
+
+class WriteReviewBinding extends Bindings {
+  @override
+  void dependencies() {
+    Get.lazyPut<WriteReviewController>(() => WriteReviewController());
+  }
+}
